+++ package.json
@@ -11,6 +11,7 @@
     "react-icons": "^5.0.1",
     "react-router-dom": "^6.22.0",
     "react-scripts": "5.0.1",
+    "react-spring-3d-carousel": "^1.3.4",
     "swiper": "^11.0.5",
     "web-vitals": "^2.1.4"
   },
