+++ src/components/footer/Footer.jsx
@@ -1,95 +1,199 @@
 import React from "react";
-import { IoMdCall, IoMdMail, IoMdPin } from "react-icons/io";
+import Logo from "../../assets/images/logo_final.png";
 
+import { IoMdCall, IoMdMail, IoMdPin } from "react-icons/io";
 const Footer = () => {
   return (
-    <section className="self-stretch bg-midnightblue flex flex-col items-center justify-start pt-[57px] pb-[116px] pr-[60px] pl-[57px] box-border gap-[46px] max-w-full shrink-0 text-left text-base text-white font-inter mq750:gap-[46px] mq750:pt-[37px] mq750:pb-[75px] mq750:pr-[30px] mq750:pl-7 mq750:box-border">
-    <div className="w-[1280px] h-[470px] relative bg-midnightblue hidden max-w-full" />
-    <footer className="flex flex-row items-center justify-start py-0 px-[3px] box-border gap-[75px] max-w-full text-left text-base text-white font-inter mq750:gap-[75px] mq1225:flex-wrap mq1225:gap-[75px]">
-      <div className="w-[535px] flex flex-col items-start justify-start gap-[36px] max-w-full mq450:gap-[36px]">
-        <div className="flex flex-row items-center justify-start gap-[7px]">
-          <img
-            className="h-[40.5px] w-[30px] relative object-cover z-[1]"
-            loading="eager"
-            alt=""
-            src={require("../../assets/images/logo_final.png")}
-            />
-          <b className="relative z-[1]">CUET-TestKnock</b>
-        </div>
-        <b className="w-[445px] h-[141px] relative inline-block whitespace-pre-wrap shrink-0 max-w-full z-[1]">
-          At TestKnock, we are dedicated to more than just academic success.
-          Our unwavering commitment extends to providing you with the
-          essential tools and expert guidance. So join us on this
-          transformative expedition and discover where educational excellence
-          is not just an aspiration but an inspiring and achievable pursuit.
-        </b>
-      </div>
-      <div className="h-[195px] flex flex-col items-start justify-start">
-        <div className="flex flex-col items-center justify-start gap-[27px]">
-          <b className="relative z-[1]">Company</b>
-          <div className="flex flex-col items-center justify-start py-0 px-1 gap-[10px]">
-            <b className="relative z-[1]">Home</b>
-            <b className="relative z-[1]">About</b>
-            <b className="relative z-[1]">Syllabus</b>
-            <b className="relative z-[1]">Courses</b>
+    <footer className="bg-[#0D0756] text-white overflow-hidden w-full ">
+      <div className="mx-auto max-w-screen-2xl ">
+        <div className="grid grid-cols-1 gap-8 lg:grid-cols-3">
+          <div className="grid justify-center ">
+            <div className=" flex justify-center gap-6 lg:justify-start text-[#063970] ">
+              <a to="/">
+                <div className="flex items-center">
+                  <div className="mr-4 mb-[-12px] ">
+                    <img
+                      src={Logo}
+                      alt="Logo"
+                      className="max-h-[60px] max-w-[60px] px-4 pb-2"
+                    />
+                  </div>
+                  <div className="text-white font-sans font-bold text-[120%] ml-[-8%]">
+                    <h2>
+                      <span>CUET</span>
+                      <span>-Tes</span>
+                      <span>tKn</span>
+                      <span>ock</span>
+                    </h2>
+                  </div>
+                </div>
+              </a>
+            </div>
+            <p className="mt-6 lg:text-justify md:text-[100%] text-[80%] hover:text-white text-center leading-relaxed text-white lg:max-w-xs ml-[20px] mr-[20px] ">
+              At TestKnock, we are dedicated to more than just academic success.
+              Our unwavering commitment extends to providing you with the
+              essential tools and expert guidance. So join us on this
+              transformative expedition and discover where educational
+              excellence is not just an aspiration but an inspiring and
+              achievable pursuit.
+            </p>
           </div>
-        </div>
-      </div>
-      <div className="w-[399px] flex flex-row items-start justify-start gap-[5px] max-w-full mq450:flex-wrap">
-        <div className="flex-1 flex flex-col items-start justify-start gap-[26px] min-w-[161px]">
-          <div className="flex flex-row items-start justify-start py-0 px-[31px]">
-            <b className="relative z-[1]">Whom we serve</b>
-          </div>
-          <div className="self-stretch flex flex-col items-end justify-start gap-[9px] text-center">
-            <div className="w-[220px] flex flex-row items-end justify-between gap-[20px] text-left">
-              <b className="relative z-[1]">School Students</b>
-              <IoMdMail />
+          <div className="grid grid-cols-2 gap-8 sm:grid-cols-3 md:grid-cols-3 lg:col-span-2">
+            <div className="text-center ">
+             
+              <ul className="mt-8 space-y-4 text-[80%] md:text-[100%] list-none">
+              <p className="md:text-[120%] mb-[40px] text-center text-[90%] text-white hover:text-white ">
+                Company
+              </p>
+                <li>
+                  <a
+                    className="text-white transition hover:text-white"
+                    to="/"
+                  >
+                    Home
+                  </a>
+                </li>
 
-            </div>
-            <div className="self-stretch flex flex-row items-center justify-between gap-[20px]">
-              <div className="flex flex-col items-center justify-start gap-[10px]">
-                <b className="relative z-[1]">College Students</b>
-                <b className="relative z-[1]">Educational Institutions</b>
-              </div>
-              <div className="flex flex-col items-start justify-start ">
-              <IoMdCall />
-              </div>
-            </div>
-            <div className="w-[216px] flex flex-row  justify-between gap-[20px]">
-              <b className="relative z-[1]">Private Schools</b>
-              <IoMdPin />
+                <li>
+                  <a
+                    className="text-white transition hover:text-white"
+                    to="/about"
+                  >
+                    About
+                  </a>
+                </li>
 
-            </div>
-            <div className="self-stretch flex flex-row items-start justify-start py-0 px-7">
-              <b className="relative z-[1]">Private Colleges</b>
-            </div>
-          </div>
-        </div>
-        <div className="w-[146px] flex flex-col items-start justify-start gap-[27px] min-w-[146px] mq450:flex-1">
-          <div className="flex flex-row items-start justify-start py-0 px-[5px]">
-            <b className="relative z-[1]">Contact Us</b>
-          </div>
-          <div className="self-stretch flex flex-col items-end justify-start gap-[31px] text-sm">
-            <div className="flex flex-col items-start justify-start gap-[25px]">
-              <b className="relative whitespace-nowrap z-[1]">
+                <li>
+                  <a
+                    className="text-white transition hover:text-white"
+                    to="/syllabus"
+                  >
+                    Syllabus
+                  </a>
+                </li>
+
                 
-                info@testknock.com
-              </b>
-              <b className="relative z-[1]">9953617456</b>
+                <li>
+                  <a
+                    className="text-white transition hover:text-white"
+                    to="/courses"
+                  >
+                    Courses
+                  </a>
+                </li>
+
+              </ul>
             </div>
-            <div className="self-stretch flex flex-row items-start justify-start py-0 pr-[3px] pl-0 text-xs">
-              <b className="flex-1 relative z-[1]">
-                Office No. - 42, Durga Vihar, near Amrapali Saphhire, Sec-45,
-                Noida-201301
-              </b>
+            <div className="text-center ">
+             
+              <ul className="mt-8 space-y-4 text-[80%] md:text-[100%] list-none">
+              <p className="md:text-[120%] mb-[40px] text-[90%] hover:text-white  text-white whitespace-nowrap">
+                Whom We Serve
+              </p>
+                <li>
+                  <p className="text-white transition hover:text-white whitespace-nowrap">
+                    School Students
+                  </p>
+                </li>
+
+                <li>
+                  <p className="text-white transition hover:text-white whitespace-nowrap">
+                    College Students
+                  </p>
+                </li>
+
+                <li>
+                  <p className="text-white  transition hover:text-white whitespace-nowrap">
+                    Educational Institutions
+                  </p>
+                </li>
+
+                <li>
+                  <p className="text-white transition  hover:text-white whitespace-nowrap">
+                    Private Schools
+                  </p>
+                </li>
+                <li>
+                  <p className="text-white transition  hover:text-white whitespace-nowrap">
+                    Private Colleges
+                  </p>
+                </li>
+              </ul>
+            </div>
+            <div className="text-center  max-sm:hidden sm:text-left">
+              
+              <ul className="mt-8 space-y-4 text-[80%] md:text-[100%]">
+              <p className="md:text-[120%] text-center items-center align-middle mb-[40px] text-[90%] hover:text-white font-medium text-white">
+                Contact Us
+              </p>
+                <li className="flex items-center hover:text-white text-white  gap-1.5 justify-start ">
+                  <a
+                    href="mailto:info@testknock.com"
+                    className="gap-1.5 flex no-underline text-white "
+                  >
+                    <IoMdMail className='relative top-[5px]'  />
+                    <h>info@testknock.com</h>
+                  </a>
+                </li>
+
+                <li className="flex items-center hover:text-white text-white  gap-1.5 justify-start ">
+                  <a href="tel:9810033495" className="gap-1.5 flex no-underline text-white ">
+                    <IoMdCall className='relative top-[5px]' />
+                    <h>9953617456</h>
+                  </a>
+                </li>
+
+                <li className="flex items-start hover:text-white text-white justify-center gap-1.5 ltr:sm:justify-start rtl:sm:justify-end">
+                  <IoMdPin />
+                  <h className="-mt-0.5 flex-1 not-italic">
+                   Office No. - 42, Durga Vihar, <br/>
+                  Near Amrapali Saphhire, <br/>
+                   Sec-45, Noida, 201301
+                  </h>
+                </li>
+              </ul>
             </div>
           </div>
         </div>
-      </div>
-    </footer>
-    <div className="mx-auto h-[34px] flex flex-col items-start justify-start py-0 px-px box-border gap-[15px] max-w-full mq750:h-auto">
-      <div className="self-stretch h-px relative box-border z-[1] border-t-[1px] border-solid border-white" />
-      <div className="w-[1050px] flex flex-row items-start justify-start py-0 px-[34px] box-border max-w-full">
+        <div className="sm:hidden block text-center mt-[6vw] ">
+          <h className="text-[110%] font-medium text-white">Contact Us</h>
+          <ul className="mt-8  space-y-4 text-[80%]">
+            <li className="flex items-center text-white justify-center gap-1.5 ">
+              <a href="mailto:info@testknock.com">
+                <h className="flex items-center text-white justify-center gap-1.5 ">
+                  <IoMdMail />
+                  <span className="flex-1 hover:text-white text-white">
+                    info@testknock.com
+                  </span>
+                </h>
+              </a>
+            </li>
+            <li className="flex items-center text-white justify-center gap-1.5 ">
+              <a href="tel:9953617456">
+                <h className="flex items-center text-white justify-center gap-1.5 ltr:sm:justify-start rtl:sm:justify-end">
+                  <IoMdCall />
+
+                  <span className="flex-1 hover:text-white text-white">
+                    9953617456
+                  </span>
+                </h>
+              </a>
+            </li>
+            <li className="flex mx-auto w-[300px] relative text-ellipsis text-white justify-center items-center">
+              <h className="flex  text-white justify-center gap-1.5 ">
+                <IoMdPin className="relative top-[3px] left-[16px]" />
+                <span className="flex-1 hover:text-white text-white">
+                  {" "}
+                  Office No. - 42, Durga Vihar, Near Amrapali Saphhire, Sec-45,
+                  Noida, 201301
+                </span>
+              </h>
+            </li>
+          </ul>
+        </div>
+        <div className="mx-auto my-[50px]  flex flex-col items-start justify-start py-0 px-px box-border gap-[15px] max-w-full ">
+      <div className="self-stretch mx-[40px]  relative box-border z-[1] border-t-[1px] border-solid border-white" />
+      <div className="w-[1050px] mx-auto flex flex-row items-start justify-start py-0 px-[34px] box-border max-w-full">
         <div className="flex-1 flex flex-row items-start justify-between gap-[20px] max-w-full mq750:flex-wrap">
           <div className="flex flex-row items-start justify-start gap-[76px] max-w-full mq750:flex-wrap mq450:gap-[76px]">
             <div className="flex flex-row items-start justify-start gap-[6px]">
@@ -107,7 +211,8 @@ const Footer = () => {
         </div>
       </div>
     </div>
-  </section>
+      </div>
+    </footer>
   );
 };
